+++ src/schemas/question.ts
@@ -1,6 +1,6 @@
 import gql from "graphql-tag";
 
-export const typedefs = gql`
+export const typeDefs = gql`
   extend type Query {
     "Get the Questions that belong to the associated sectionId"
     questions(sectionId: Int!): [Question]
@@ -10,11 +10,13 @@ export const typedefs = gql`
 
 extend type Mutation {
     "Create a new Question"
-    addQuestion(input: AddQuestionInput): Question
+    addQuestion(input: AddQuestionInput!): Question!
     "Update a Question"
-    updateQuestion(input: UpdateQuestionInput): Question
+    updateQuestion(input: UpdateQuestionInput!): Question!
     "Delete a Question"
-    removeQuestion(questionId: Int!): Boolean
+    removeQuestion(questionId: Int!): Question
+    "Separate Question update specifically for options"
+    updateQuestionOptions(questionId: Int!, required:Boolean = false ): Question
   }
 
 "Question always belongs to a Section, which always belongs to a Template"