+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ### Added
+- Built Question resolvers and models(#13)
 - Fixed some bugs to allow frontend to access token change(Frontend #116)
 - Added data migrations for QuestionType, Question, QuestionCondition, VersionedQuestion and VersionedQuestionCondition
 - Added missing VersionedQuestionCondition schema file